deploy 1

+++ app/Http/Controllers/OffreController.php
@@ -2,9 +2,11 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Middleware\VerifLicence;
 use App\Repositories\OffreRepository;
 use App\Repositories\SalleRepository;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class OffreController extends Controller
 {
@@ -13,6 +15,8 @@ class OffreController extends Controller
 
     public function __construct(OffreRepository $offreRepository,SalleRepository $salleRepository)
     {
+        $this->middleware([VerifLicence::class])->except(['index',"show"]);
+
         $this->offreRepository = $offreRepository;
         $this->salleRepository = $salleRepository;
     }
@@ -23,7 +27,8 @@ class OffreController extends Controller
      */
     public function index()
     {
-        $offres = $this->offreRepository->getAll();
+        $user = Auth::user();
+        $offres = $this->offreRepository->getOffreBySalle($user->salle_id);
         return view('offre.index',compact('offres'));
     }
 
@@ -45,12 +50,22 @@ class OffreController extends Controller
      */
     public function store(Request $request)
     {
+
+        $user = auth()->user();
+        $salle = $user->salle;
+
+        if (!$salle->hasActiveLicence()) {
+            return redirect()->back()->withErrors(['licence' => 'La salle n\'a pas de licence active.']);
+        }
+
         $request->validate([
 
             'nom' => 'required|string',
+            'duree' => 'required|integer',
+            'prix' => 'required|integer',
 
         ]);
-
+        $request->merge(["salle_id"=>Auth::user()->salle_id]);
         $offre = $this->offreRepository->store($request->all());
         return redirect('offre');
 
@@ -89,6 +104,13 @@ class OffreController extends Controller
      */
     public function update(Request $request, $id)
     {
+         $user = auth()->user();
+        $salle = $user->salle;
+
+        if (!$salle->hasActiveLicence()) {
+            return redirect()->back()->withErrors(['licence' => 'La salle n\'a pas de licence active.']);
+        }
+
         $this->offreRepository->update($id, $request->all());
          return redirect('offre');
     }
@@ -101,11 +123,22 @@ class OffreController extends Controller
      */
     public function destroy($id)
     {
+         $user = auth()->user();
+        $salle = $user->salle;
+
+        if (!$salle->hasActiveLicence()) {
+            return redirect()->back()->withErrors(['licence' => 'La salle n\'a pas de licence active.']);
+        }
+
         $this->offreRepository->destroy($id);
         return redirect('offre');
     }
         //
 
-
+    public function updateOffreByEtatAndSalle($id,$etat)
+    {
+        $this->offreRepository->updateOffreByEtatAndSalle($id,$etat);
+        return redirect('offre');
+    }
 
 }
